refactor : Post 클래스 post/domain 폴더로 이동
