+++ src/main/java/com/kakaotech/team14backend/inner/post/repository/PostRepository.java
@@ -1,6 +1,6 @@
 package com.kakaotech.team14backend.inner.post.repository;
 
-import com.kakaotech.team14backend.inner.post.model.Post;
+import com.kakaotech.team14backend.post.domain.Post;
 import com.kakaotech.team14backend.outer.post.dto.GetIncompletePopularPostDTO;
 import io.lettuce.core.dynamic.annotation.Param;
 import java.util.List;
