+++ src/main/java/com/kakaotech/team14backend/inner/post/usecase/FindMyPostUsecase.java
@@ -1,14 +1,10 @@
 package com.kakaotech.team14backend.inner.post.usecase;
 
-import com.kakaotech.team14backend.inner.post.model.Post;
-import com.kakaotech.team14backend.inner.post.model.PostLike;
-import com.kakaotech.team14backend.inner.post.repository.PostLikeRepository;
+import com.kakaotech.team14backend.post.domain.Post;
 import com.kakaotech.team14backend.inner.post.repository.PostRepository;
 import com.kakaotech.team14backend.outer.post.dto.GetMyPostResponseDTO;
 import com.kakaotech.team14backend.outer.post.mapper.PostMapper;
 import com.kakaotech.team14backend.outer.post.service.FindLikeStatusService;
-import java.util.Optional;
-import javax.swing.text.StyledEditorKit.BoldAction;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
