+++ src/main/java/com/kakaotech/team14backend/inner/post/usecase/UpdatePostViewCountUsecase.java
@@ -3,7 +3,7 @@ package com.kakaotech.team14backend.inner.post.usecase;
 import com.kakaotech.team14backend.common.RedisKey;
 import com.kakaotech.team14backend.common.ScanRedisKey;
 import com.kakaotech.team14backend.exception.PostNotFoundException;
-import com.kakaotech.team14backend.inner.post.model.Post;
+import com.kakaotech.team14backend.post.domain.Post;
 import com.kakaotech.team14backend.inner.post.repository.PostRepository;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
