+++ src/test/java/com/kakaotech/team14backend/inner/post/model/PostTest.java
@@ -1,5 +1,6 @@
 package com.kakaotech.team14backend.inner.post.model;
 
+import com.kakaotech.team14backend.post.domain.Post;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import static org.assertj.core.api.Assertions.assertThat;
