+++ src/test/java/com/kakaotech/team14backend/inner/post/usecase/SaveTemporaryPopularPostListUsecaseTest.java
@@ -1,7 +1,7 @@
 package com.kakaotech.team14backend.inner.post.usecase;
 
 import com.kakaotech.team14backend.common.RedisKey;
-import com.kakaotech.team14backend.inner.post.model.Post;
+import com.kakaotech.team14backend.post.domain.Post;
 import com.kakaotech.team14backend.inner.post.repository.PostRepository;
 import java.util.List;
 import java.util.Set;
